time duration change to long

+++ tgd/src/main/java/com/jiaxy/tgd/FailFastTokenBucketLimiter.java
@@ -25,7 +25,7 @@ public class FailFastTokenBucketLimiter extends RateLimiter {
     }
 
     @Override
-    public void syncAvailableToken(double nowMicros) {
+    public void syncAvailableToken(long nowMicros) {
         if (nowMicros > nextGenTokenMicros){
             double newTokens = (nowMicros - nextGenTokenMicros) / stableIntervalTokenMicros;
             availableToken = Math.min(maxToken,availableToken + newTokens);
@@ -35,7 +35,7 @@ public class FailFastTokenBucketLimiter extends RateLimiter {
 
     @Override
     public double getToken(double requiredToken) {
-        double nowMicros = duration();
+        long nowMicros = duration();
         synchronized (mutex){
             syncAvailableToken(nowMicros);
             double tokenPermitted = Math.min(requiredToken,availableToken);
